chore: split out backup/restore into separate component
