+++ packages/server-admin-ui/src/views/ServerConfig/Settings.js
@@ -18,16 +18,15 @@ import {
   Table,
   Progress
 } from 'reactstrap'
-import escape from 'escape-html'
-import { restart } from '../../actions'
+
+import VesselConfiguration from './VesselConfiguration'
+import LogFiles from './Logging'
+import BackupRestore from './BackupRestore'
 
 const RESTORE_NONE = 0
 const RESTORE_VALIDATING = 1
 const RESTORE_CONFIRM = 2
 const RESTORE_RUNNING = 3
-
-import VesselConfiguration from './VesselConfiguration'
-import LogFiles from './Logging'
 
 function fetchSettings () {
   fetch(`/settings`, {
@@ -44,123 +43,18 @@ class ServerSettings extends Component {
     super(props)
     this.state = {
       hasData: false,
-      restoreFile: null,
-      restoreState: RESTORE_NONE
     }
     this.fetchSettings = fetchSettings.bind(this)
     this.handleChange = this.handleChange.bind(this)
-    this.fileChanged = this.fileChanged.bind(this)
     this.handleOptionChange = this.handleOptionChange.bind(this)
     this.handleInterfaceChange = this.handleInterfaceChange.bind(this)
     this.handleSaveSettings = this.handleSaveSettings.bind(this)
-    this.handleRestoreFileChange = this.handleRestoreFileChange.bind(this)
-    this.backup = this.backup.bind(this)
-    this.validate = this.validate.bind(this)
-    this.restore = this.restore.bind(this)
-    this.restart = this.restart.bind(this)
   }
 
   componentDidMount () {
     this.fetchSettings()
   }
-
-  fileChanged (event) {
-    this.setState({ ...this.state,
-      restoreFile: event.target.files[0]
-    })
-  }
-
-  backup() {
-    window.location = '/backup'
-  }
-
-  restore() {
-    console.log(this.state.restoreContents)
-    fetch(`/restore`, {
-      credentials: 'include',
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(this.state.restoreContents)
-    })
-      .then(response => {
-        if (!response.ok) {
-          return response.text()
-        }
-      })
-      .then(res => {
-        if ( typeof res === 'string' ) {
-          alert(res)
-          this.setState({restoreState: RESTORE_NONE, restoreFile:null})
-
-        } else {
-          this.setState({restoreState: RESTORE_RUNNING})
-        }
-      })
-      .catch(error => {
-        alert(error.message)
-      })
-  }
-
-  restart() {
-    this.props.restart()
-    this.setState({restoreState: RESTORE_NONE})
-    window.location = "/admin/#/dashboard"
-  }
-
-  validate() {
-    if ( !this.state.restoreFile ) {
-      alert('Please choose a file')
-      return
-    }
-
-    const data = new FormData() 
-    data.append('file', this.state.restoreFile)
-    
-    this.setState({restoreState: RESTORE_VALIDATING})
-    fetch(`/validateBackup`, {
-      credentials: 'include',
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json'
-      },
-      body: data
-    })
-      .then(response => {
-        if (response.ok) {
-          return response.json()
-        } else {
-          return response.text()
-        }
-      })
-      .then(res => {
-        if ( typeof res === 'string' ) {
-          alert(res)
-          this.setState({restoreState: RESTORE_NONE, restoreFile:null})
-
-        } else {
-          const restoreContents = {}
-          res.forEach(filename => {
-            restoreContents[filename] = true
-          })
-          this.setState({restoreState: RESTORE_CONFIRM, restoreContents})
-        }
-      })
-      .catch(error => {
-        alert(error.message)
-      })
-  }
-
-  handleRestoreFileChange (event) {
-    const value =
-      event.target.type === 'checkbox'
-        ? event.target.checked
-        : event.target.value
-    this.state.restoreContents[event.target.name] = value
-    this.setState({ restoreContents: this.state.restoreContents })
-  }
-  
+ 
   handleChange (event) {
     const value =
       event.target.type === 'checkbox'
@@ -207,7 +101,6 @@ class ServerSettings extends Component {
     return (
       this.state.hasData && (
           <div className='animated fadeIn'>
-          {this.state.restoreState === RESTORE_NONE && !this.props.restoreStatus.state && (
           <Card>
             <CardHeader>
               <i className='fa fa-align-justify' /><strong>Server Settings</strong>
@@ -387,163 +280,27 @@ class ServerSettings extends Component {
               </Badge>
             </CardFooter>
           </Card>
-          )}
-          <Card>
-            <CardHeader>Backup and Restore</CardHeader>
-            <CardBody>
-              <Form
-                action=''
-                method='post'
-                encType='multipart/form-data'
-                className='form-horizontal'
-              >
-            {this.state.restoreState === RESTORE_NONE && !this.props.restoreStatus.state && (
-            <div>
-            <FormText color='muted'>
-              The backup will contain the server and plugin settings only.
-            </FormText><br/>
-              <FormGroup row>
-                  <Col xs='12' md={fieldColWidthMd}>
-                    <Input
-                      type='file'
-                      name='backupFile'
-                      onChange={this.fileChanged}
-                    />
-                    <FormText color='muted'>
-                      Your existing settings will be overwritten
-                    </FormText>
-                  </Col>
-              </FormGroup>
-              </div>
-             )}
-             {this.state.restoreState === RESTORE_CONFIRM && (
-                  <FormGroup check>
-                  <Col xs='12' md={fieldColWidthMd}>
-                      {_.keys(this.state.restoreContents).map(name => {
-                        return (
-                          <div key={name}>
-                            <Label className='switch switch-text switch-primary'>
-                              <Input
-                                type='checkbox'
-                                id={name}
-                                name={name}
-                                className='switch-input'
-                                onChange={this.handleRestoreFileChange}
-                                checked={this.state.restoreContents[name]}
-                              />
-                              <span
-                                className='switch-label'
-                                data-on='Yes'
-                                data-off='No'
-                              />
-                              <span className='switch-handle' />
-                            </Label>{' '}
-                            {name}
-                          </div>
-                        )
-                      })}
-                  </Col>
-                  </FormGroup>
-             )}
-             {this.props.restoreStatus && this.props.restoreStatus.state && this.props.restoreStatus.state !== 'Complete' && (
-             <div>
-             <FormGroup row>             
-               <Col xs='12' md={fieldColWidthMd}>
-               <FormText>{this.props.restoreStatus.state} : {this.props.restoreStatus.message}</FormText>
-               </Col>
-             </FormGroup>
-             <FormGroup row>             
-               <Col xs='12' md={fieldColWidthMd}>
-               <Progress
-                 animated
-                 color='success'
-                 value={this.props.restoreStatus.percentComplete}
-               />
-               </Col>
-             </FormGroup>
-             </div>
-             )}
-             {this.props.restoreStatus.state && this.props.restoreStatus.state === 'Complete' && (
-             <div>
-             <FormGroup row>             
-               <Col xs='12' md={fieldColWidthMd}>
-               <FormText>Please Restart</FormText>
-               </Col>
-             </FormGroup>
-             <FormGroup row>             
-               <Col xs='12' md={fieldColWidthMd}>
-                 <Button
-                    size='sm'
-                    color='danger'
-                    onClick={this.restart}
-                  >
-                   {this.props.restarting ? (
-                     <i className='fa fa-circle-o-notch fa-spin' />
-                   ) : (
-                     <i className='fa fa-circle-o-notch' /> 
-                   )} Restart
-                  </Button>
-               </Col>
-             </FormGroup>
-             </div>
-             )}
-             </Form>
-            </CardBody>
-            <CardFooter>
-            {this.state.restoreState === RESTORE_NONE && !this.props.restoreStatus.state && (
-            <div>
-              <Button
-                size='sm'
-                color='primary'
-                onClick={this.backup}
-              >
-                <i className='fa fa-dot-circle-o' /> Backup
-              </Button>{' '}
-              <Button
-                size='sm'
-                color='danger'
-                onClick={this.validate}
-              >
-                <i className='fa fa-dot-circle-o' /> Restore
-              </Button>{' '}
-            </div>
-            )}
-            {this.state.restoreState === RESTORE_CONFIRM && (
-            <div>
-              <Button
-                size='sm'
-                color='primary'
-                onClick={this.cancelRestore}
-              >
-                <i className='fa fa-dot-circle-o' /> Cancel
-              </Button>{' '}
-              <Button
-                size='sm'
-                color='danger'
-                onClick={this.restore}
-              >
-                <i className='fa fa-dot-circle-o' /> Confirm
-              </Button>
-              </div>
-            )}
-            </CardFooter>
-          </Card>
         </div>
       )
     )
   }
 }
 
-const ReduxedSettings = connect(({restoreStatus, restarting}) => ({restoreStatus, restarting}), {restart})(ServerSettings)
+const ReduxedSettings = connect()(ServerSettings)
 
 class Settings extends Component {
   render() {
     return <div>
-      <VesselConfiguration/>
-      <ReduxedSettings/>
-      <LogFiles/>
+      {!this.props.restoreStatus.state && (
+        <div>
+        <VesselConfiguration/>
+        <ReduxedSettings/>
+        <LogFiles/>
+        </div>
+      )}
+      <BackupRestore/>
       </div>
   }
 }
 
-export default Settings
+export default connect(({restoreStatus}) => ({restoreStatus}))(Settings)
